+++ SEPADirectDebit/src/main/java/com/braintreepayments/api/SEPADirectDebitLauncher.java
@@ -21,7 +21,7 @@ public class SEPADirectDebitLauncher {
      * Used to launch the SEPA mandate in a web browser and deliver results to your Activity
      *
      * @param callback a {@link SEPADirectDebitLauncherCallback} to handle the result of
-     * {@link SEPADirectDebitLauncher#launch(FragmentActivity, SEPADirectDebitResponse)}
+     * {@link SEPADirectDebitLauncher#launch(FragmentActivity, SEPADirectDebitPaymentAuthRequest)}
      */
     public SEPADirectDebitLauncher(@NonNull SEPADirectDebitLauncherCallback callback) {
         this(new BrowserSwitchClient(), callback);
@@ -40,14 +40,15 @@ public class SEPADirectDebitLauncher {
      * {@link SEPADirectDebitLauncher#SEPADirectDebitLauncher(SEPADirectDebitLauncherCallback)}
      *
      * @param activity       an Android {@link FragmentActivity}
-     * @param sepaDirectDebitResponse the result of the SEPA mandate received from invoking
-     *                       {@link SEPADirectDebitClient#tokenize(SEPADirectDebitRequest, SEPADirectDebitFlowStartedCallback)}
+     * @param paymentAuthRequest the result of the SEPA mandate received from invoking
+     *                       {@link SEPADirectDebitClient#createPaymentAuthRequest(SEPADirectDebitRequest, SEPADirectDebitPaymentAuthRequestCallback)}
      */
-    public void launch(@NonNull FragmentActivity activity, @NonNull SEPADirectDebitResponse sepaDirectDebitResponse) {
+    public void launch(@NonNull FragmentActivity activity, @NonNull
+    SEPADirectDebitPaymentAuthRequest paymentAuthRequest) {
         try {
-            browserSwitchClient.start(activity, sepaDirectDebitResponse.getBrowserSwitchOptions());
+            browserSwitchClient.start(activity, paymentAuthRequest.getBrowserSwitchOptions());
         } catch (BrowserSwitchException e) {
-            callback.onResult(new SEPADirectDebitBrowserSwitchResult(e));
+            callback.onResult(new SEPADirectDebitPaymentAuthResult(e));
         }
     }
 
@@ -55,16 +56,16 @@ public class SEPADirectDebitLauncher {
      * Captures and delivers the result of the browser-based SEPA mandate flow.
      * <p>
      * For most integrations, this method should be invoked in the onResume method of the Activity
-     * used to invoke {@link SEPADirectDebitLauncher#launch(FragmentActivity, SEPADirectDebitResponse)}.
+     * used to invoke {@link SEPADirectDebitLauncher#launch(FragmentActivity, SEPADirectDebitPaymentAuthRequest)}.
      * <p>
      * If the Activity used to launch the SEPA mandate is configured with
      * android:launchMode="singleTop", this method should be invoked in the onNewIntent method of
      * the Activity, after invoking setIntent(intent).
      * <p>
-     * This method will deliver a {@link SEPADirectDebitBrowserSwitchResult} to the
+     * This method will deliver a {@link SEPADirectDebitPaymentAuthResult} to the
      * {@link SEPADirectDebitLauncherCallback} used to instantiate this class. The
-     * {@link SEPADirectDebitBrowserSwitchResult} should be passed to
-     * {@link SEPADirectDebitClient#onBrowserSwitchResult(SEPADirectDebitBrowserSwitchResult, SEPADirectDebitBrowserSwitchResultCallback)}
+     * {@link SEPADirectDebitPaymentAuthResult} should be passed to
+     * {@link SEPADirectDebitClient#tokenize(SEPADirectDebitPaymentAuthResult, SEPADirectDebitTokenizeCallback)}
      *
      * @param context the context used to check for pending results
      * @param intent  the intent to return to your application containing a deep link result from
@@ -73,7 +74,7 @@ public class SEPADirectDebitLauncher {
     public void handleReturnToAppFromBrowser(@NonNull Context context, @NonNull Intent intent) {
         BrowserSwitchResult result = browserSwitchClient.parseResult(context, SEPA_DEBIT, intent);
         if (result != null) {
-            callback.onResult(new SEPADirectDebitBrowserSwitchResult(result));
+            callback.onResult(new SEPADirectDebitPaymentAuthResult(result));
             browserSwitchClient.clearActiveRequests(context);
         }
     }