+++ SEPADirectDebit/src/test/java/com/braintreepayments/api/SEPADirectDebitClientUnitTest.java
@@ -6,7 +6,6 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.isNull;
 import static org.mockito.Mockito.mock;
@@ -32,8 +31,8 @@ public class SEPADirectDebitClientUnitTest {
     private BraintreeClient braintreeClient;
     private CreateMandateResult createMandateResult;
     private SEPADirectDebitRequest sepaDirectDebitRequest;
-    private SEPADirectDebitBrowserSwitchResultCallback sepaBrowserSwitchResultCallback;
-    private SEPADirectDebitFlowStartedCallback sepaFlowStartedCallback;
+    private SEPADirectDebitTokenizeCallback sepaTokenizeCallback;
+    private SEPADirectDebitPaymentAuthRequestCallback sepaFlowStartedCallback;
 
     @Before
     public void beforeEach() {
@@ -50,8 +49,8 @@ public class SEPADirectDebitClientUnitTest {
         );
         sepaDirectDebitRequest = new SEPADirectDebitRequest();
 
-        sepaBrowserSwitchResultCallback = mock(SEPADirectDebitBrowserSwitchResultCallback.class);
-        sepaFlowStartedCallback = mock(SEPADirectDebitFlowStartedCallback.class);
+        sepaTokenizeCallback = mock(SEPADirectDebitTokenizeCallback.class);
+        sepaFlowStartedCallback = mock(SEPADirectDebitPaymentAuthRequestCallback.class);
     }
 
     @Test
@@ -64,14 +63,14 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        sut.tokenize(sepaDirectDebitRequest, sepaFlowStartedCallback);
+        sut.createPaymentAuthRequest(sepaDirectDebitRequest, sepaFlowStartedCallback);
         verify(braintreeClient).sendAnalyticsEvent("sepa-direct-debit.selected.started");
 
-        ArgumentCaptor<SEPADirectDebitResponse> captor =
-                ArgumentCaptor.forClass(SEPADirectDebitResponse.class);
+        ArgumentCaptor<SEPADirectDebitPaymentAuthRequest> captor =
+                ArgumentCaptor.forClass(SEPADirectDebitPaymentAuthRequest.class);
         verify(sepaFlowStartedCallback).onResult(captor.capture(), isNull());
 
-        SEPADirectDebitResponse sepaResponseResult = captor.getValue();
+        SEPADirectDebitPaymentAuthRequest sepaResponseResult = captor.getValue();
 
         verify(braintreeClient).sendAnalyticsEvent("sepa-direct-debit.create-mandate.requested");
         verify(braintreeClient).sendAnalyticsEvent("sepa-direct-debit.create-mandate.success");
@@ -109,9 +108,10 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        sut.tokenize(sepaDirectDebitRequest, sepaFlowStartedCallback);
+        sut.createPaymentAuthRequest(sepaDirectDebitRequest, sepaFlowStartedCallback);
 
-        ArgumentCaptor<SEPADirectDebitResponse> captor = ArgumentCaptor.forClass(SEPADirectDebitResponse.class);
+        ArgumentCaptor<SEPADirectDebitPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                SEPADirectDebitPaymentAuthRequest.class);
         verify(sepaFlowStartedCallback).onResult(captor.capture(), isNull());
         assertEquals(captor.getValue().getNonce(), nonce);
         verify(braintreeClient).sendAnalyticsEvent("sepa-direct-debit.tokenize.success");
@@ -137,7 +137,7 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        sut.tokenize(sepaDirectDebitRequest, sepaFlowStartedCallback);
+        sut.createPaymentAuthRequest(sepaDirectDebitRequest, sepaFlowStartedCallback);
         verify(sepaFlowStartedCallback).onResult(isNull(), eq(exception));
         verify(braintreeClient).sendAnalyticsEvent("sepa-direct-debit.tokenize.failure");
     }
@@ -159,7 +159,7 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        sut.tokenize(sepaDirectDebitRequest, sepaFlowStartedCallback);
+        sut.createPaymentAuthRequest(sepaDirectDebitRequest, sepaFlowStartedCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
         verify(sepaFlowStartedCallback).onResult(isNull(), captor.capture());
@@ -186,7 +186,7 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        sut.tokenize(sepaDirectDebitRequest, sepaFlowStartedCallback);
+        sut.createPaymentAuthRequest(sepaDirectDebitRequest, sepaFlowStartedCallback);
         verify(braintreeClient, never()).startBrowserSwitch(any(FragmentActivity.class),
                 any(BrowserSwitchOptions.class));
         verify(braintreeClient).sendAnalyticsEvent("sepa-direct-debit.create-mandate.success");
@@ -206,7 +206,7 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        sut.tokenize(sepaDirectDebitRequest, sepaFlowStartedCallback);
+        sut.createPaymentAuthRequest(sepaDirectDebitRequest, sepaFlowStartedCallback);
         verify(sepaFlowStartedCallback).onResult(isNull(), eq(error));
         verify(braintreeClient).sendAnalyticsEvent("sepa-direct-debit.create-mandate.failure");
     }
@@ -221,11 +221,12 @@ public class SEPADirectDebitClientUnitTest {
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
         Exception expectedError = new Exception("error");
-        SEPADirectDebitBrowserSwitchResult payPalBrowserSwitchResult = new SEPADirectDebitBrowserSwitchResult(expectedError);
-        sut.onBrowserSwitchResult(payPalBrowserSwitchResult, sepaBrowserSwitchResultCallback);
+        SEPADirectDebitPaymentAuthResult
+                payPalBrowserSwitchResult = new SEPADirectDebitPaymentAuthResult(expectedError);
+        sut.tokenize(payPalBrowserSwitchResult, sepaTokenizeCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
-        verify(sepaBrowserSwitchResultCallback).onResult(isNull(), captor.capture());
+        verify(sepaTokenizeCallback).onResult(isNull(), captor.capture());
 
         assertNotNull(captor.getValue());
         assertEquals(expectedError, captor.getValue());
@@ -240,11 +241,12 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        SEPADirectDebitBrowserSwitchResult payPalBrowserSwitchResult = new SEPADirectDebitBrowserSwitchResult((BrowserSwitchResult) null);
-        sut.onBrowserSwitchResult(payPalBrowserSwitchResult, sepaBrowserSwitchResultCallback);
+        SEPADirectDebitPaymentAuthResult
+                payPalBrowserSwitchResult = new SEPADirectDebitPaymentAuthResult((BrowserSwitchResult) null);
+        sut.tokenize(payPalBrowserSwitchResult, sepaTokenizeCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
-        verify(sepaBrowserSwitchResultCallback).onResult(isNull(), captor.capture());
+        verify(sepaTokenizeCallback).onResult(isNull(), captor.capture());
 
         assertNotNull(captor.getValue());
         assertEquals("An unexpected error occurred.", captor.getValue().getMessage());
@@ -263,12 +265,13 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        SEPADirectDebitBrowserSwitchResult sepaBrowserSwitchResult = new SEPADirectDebitBrowserSwitchResult(browserSwitchResult);
+        SEPADirectDebitPaymentAuthResult
+                sepaBrowserSwitchResult = new SEPADirectDebitPaymentAuthResult(browserSwitchResult);
 
-        sut.onBrowserSwitchResult(sepaBrowserSwitchResult, sepaBrowserSwitchResultCallback);
+        sut.tokenize(sepaBrowserSwitchResult, sepaTokenizeCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
-        verify(sepaBrowserSwitchResultCallback).onResult(isNull(), captor.capture());
+        verify(sepaTokenizeCallback).onResult(isNull(), captor.capture());
         verify(braintreeClient).sendAnalyticsEvent("sepa-direct-debit.browser-switch.canceled");
 
         Exception exception = captor.getValue();
@@ -300,9 +303,10 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        SEPADirectDebitBrowserSwitchResult sepaBrowserSwitchResult = new SEPADirectDebitBrowserSwitchResult(browserSwitchResult);
+        SEPADirectDebitPaymentAuthResult
+                sepaBrowserSwitchResult = new SEPADirectDebitPaymentAuthResult(browserSwitchResult);
 
-        sut.onBrowserSwitchResult(sepaBrowserSwitchResult, sepaBrowserSwitchResultCallback);
+        sut.tokenize(sepaBrowserSwitchResult, sepaTokenizeCallback);
 
         verify(braintreeClient).sendAnalyticsEvent("sepa-direct-debit.browser-switch.success");
         verify(sepaDirectDebitApi).tokenize(eq("1234"), eq("customer-id"),
@@ -339,11 +343,12 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        SEPADirectDebitBrowserSwitchResult sepaBrowserSwitchResult = new SEPADirectDebitBrowserSwitchResult(browserSwitchResult);
+        SEPADirectDebitPaymentAuthResult
+                sepaBrowserSwitchResult = new SEPADirectDebitPaymentAuthResult(browserSwitchResult);
 
-        sut.onBrowserSwitchResult(sepaBrowserSwitchResult, sepaBrowserSwitchResultCallback);
+        sut.tokenize(sepaBrowserSwitchResult, sepaTokenizeCallback);
 
-        verify(sepaBrowserSwitchResultCallback).onResult(eq(nonce), isNull());
+        verify(sepaTokenizeCallback).onResult(eq(nonce), isNull());
         verify(braintreeClient).sendAnalyticsEvent("sepa-direct-debit.tokenize.success");
     }
 
@@ -374,11 +379,12 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        SEPADirectDebitBrowserSwitchResult sepaBrowserSwitchResult = new SEPADirectDebitBrowserSwitchResult(browserSwitchResult);
+        SEPADirectDebitPaymentAuthResult
+                sepaBrowserSwitchResult = new SEPADirectDebitPaymentAuthResult(browserSwitchResult);
 
-        sut.onBrowserSwitchResult(sepaBrowserSwitchResult, sepaBrowserSwitchResultCallback);
+        sut.tokenize(sepaBrowserSwitchResult, sepaTokenizeCallback);
 
-        verify(sepaBrowserSwitchResultCallback).onResult(isNull(), eq(exception));
+        verify(sepaTokenizeCallback).onResult(isNull(), eq(exception));
         verify(braintreeClient).sendAnalyticsEvent("sepa-direct-debit.tokenize.failure");
         verify(sepaDirectDebitApi).tokenize(eq("1234"), eq("customer-id"),
                 eq("bank-reference-token"), eq("ONE_OFF"),
@@ -413,11 +419,12 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        SEPADirectDebitBrowserSwitchResult sepaBrowserSwitchResult = new SEPADirectDebitBrowserSwitchResult(browserSwitchResult);
+        SEPADirectDebitPaymentAuthResult
+                sepaBrowserSwitchResult = new SEPADirectDebitPaymentAuthResult(browserSwitchResult);
 
-        sut.onBrowserSwitchResult(sepaBrowserSwitchResult, sepaBrowserSwitchResultCallback);
+        sut.tokenize(sepaBrowserSwitchResult, sepaTokenizeCallback);
 
-        verify(sepaBrowserSwitchResultCallback).onResult(eq(nonce), isNull());
+        verify(sepaTokenizeCallback).onResult(eq(nonce), isNull());
     }
 
     @Test
@@ -436,12 +443,13 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        SEPADirectDebitBrowserSwitchResult sepaBrowserSwitchResult = new SEPADirectDebitBrowserSwitchResult(browserSwitchResult);
+        SEPADirectDebitPaymentAuthResult
+                sepaBrowserSwitchResult = new SEPADirectDebitPaymentAuthResult(browserSwitchResult);
 
-        sut.onBrowserSwitchResult(sepaBrowserSwitchResult, sepaBrowserSwitchResultCallback);
+        sut.tokenize(sepaBrowserSwitchResult, sepaTokenizeCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
-        verify(sepaBrowserSwitchResultCallback).onResult(isNull(), captor.capture());
+        verify(sepaTokenizeCallback).onResult(isNull(), captor.capture());
         verify(braintreeClient).sendAnalyticsEvent("sepa-direct-debit.browser-switch.failure");
 
         Exception exception = captor.getValue();
@@ -464,12 +472,13 @@ public class SEPADirectDebitClientUnitTest {
         SEPADirectDebitClient sut =
                 new SEPADirectDebitClient(braintreeClient, sepaDirectDebitApi);
 
-        SEPADirectDebitBrowserSwitchResult sepaBrowserSwitchResult = new SEPADirectDebitBrowserSwitchResult(browserSwitchResult);
+        SEPADirectDebitPaymentAuthResult
+                sepaBrowserSwitchResult = new SEPADirectDebitPaymentAuthResult(browserSwitchResult);
 
-        sut.onBrowserSwitchResult(sepaBrowserSwitchResult, sepaBrowserSwitchResultCallback);
+        sut.tokenize(sepaBrowserSwitchResult, sepaTokenizeCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
-        verify(sepaBrowserSwitchResultCallback).onResult(isNull(), captor.capture());
+        verify(sepaTokenizeCallback).onResult(isNull(), captor.capture());
         verify(braintreeClient).sendAnalyticsEvent("sepa-direct-debit.browser-switch.failure");
 
         Exception exception = captor.getValue();