+++ CHANGELOG.md
@@ -67,10 +67,14 @@
     * Move `ThreeDSecureInfo` to `three-d-secure` module
   * SEPA Direct Debit
     * Remove `SEPADirectDebitLifecycleObserver` and `SEPADirectDebitListener`
-    * Add `SEPADirectDebitLauncher`, `SEPADirectDebitLauncherCallback`, `SEPADirectDebitFlowStartedCallback`,
-      `SEPADirectDebitResponse`, `SEPADirectDebitBrowserSwitchResult`, and `SEPADirectDebitBrowserSwitchResultCallback`
+    * Add `SEPADirectDebitLauncher`, `SEPADirectDebitLauncherCallback`, 
+      `SEPADirectDebitPaymentAuthRequestCallback`, `SEPADirectDebitPaymentAuthRequest`, 
+      `SEPADirectDebitPaymentAuthResult`, and `SEPADirectDebitTokenizeCallback`
     * Remove Fragment or Activity requirement from `SEPADirectDebitClient` constructor
-    * Modify `SEPADirectDebitClient#tokenize` parameters
+    * Replace `SEPADirectDebitClient#onBrowserSwitchResult` with `SEPADirectDebitClient#tokenize` and 
+      modify parameters
+    * Replace `SEPADirectDebitClient#tokenize` with`SEPADirectDebitClient#createPaymentAuthRequest` 
+      and modify parameters
 
 ## unreleased